fix(collection-edit): fix issues with update_at and update_by fields
